+++ compass.h
@@ -2,6 +2,6 @@
 #define COMPASS_H
 
 int compass_initializer();
-int get_angle();
+double get_angle();
 
 #endif
